quantization/requirements.txt
tensorflow==2.18.*
tensorflow-hub==0.16.1
pillow
silence-tensorflow
h5py
tf-keras==2.18.0
numpy
tensorflow_model_optimization